[EasyPagination] Remove PSR7 support
