+++ packages/EasyPagination/src/Bridge/Laravel/Providers/AbstractStartSizeEasyPaginationProvider.php
@@ -10,7 +10,6 @@ use EonX\EasyPagination\Interfaces\StartSizeDataFactoryInterface;
 use EonX\EasyPagination\Interfaces\StartSizeDataInterface;
 use EonX\EasyPagination\Interfaces\StartSizeDataResolverInterface;
 use EonX\EasyPagination\Resolvers\Config\StartSizeConfig;
-use EonX\EasyPsr7Factory\Interfaces\EasyPsr7FactoryInterface;
 use Illuminate\Http\Request;
 use Illuminate\Support\ServiceProvider;
 
@@ -59,12 +58,7 @@ abstract class AbstractStartSizeEasyPaginationProvider extends ServiceProvider
     private function getDataClosure(): Closure
     {
         return function (): StartSizeDataInterface {
-            /** @var \EonX\EasyPsr7Factory\Interfaces\EasyPsr7FactoryInterface $psr7Factory */
-            $psr7Factory = $this->app->get(EasyPsr7FactoryInterface::class);
-
-            return $this->app->get(StartSizeDataResolverInterface::class)->resolve(
-                $psr7Factory->createRequest($this->getRequest())
-            );
+            return $this->app->get(StartSizeDataResolverInterface::class)->resolve($this->getRequest());
         };
     }
 
@@ -73,6 +67,7 @@ abstract class AbstractStartSizeEasyPaginationProvider extends ServiceProvider
         if ($this->app->runningInConsole() === true) {
             /**
              * When running in console, a request instance is created and bound to `request` alias.
+             *
              * @see \Laravel\Lumen\Console\Kernel::setRequestForConsole
              */
             return $this->app->get('request');
